january_pnl/requirements.txt
streamlit 
pandas
plotly.express 